chore: added suggeted change to rather use a const for projectDir

+++ taqueria-plugin-ligo/compile.ts
@@ -15,10 +15,10 @@ const getInputFilename = (opts: Opts) => (sourceFile: string) => {
 }
 
 const getCompileCommand = (opts: Opts) => (sourceFile: string) => {
-    let {projectDir} = opts
-    if (process.env.PROJECT_DIR && process.env.PROJECT_DIR?.length > 0) {
-        projectDir = process.env.PROJECT_DIR
-    }
+    const projectDir = process.env.PROJECT_DIR ??  opts.projectDir
+
+    if (!projectDir) throw `No project directory provided`
+    
     const inputFile = getInputFilename (opts) (sourceFile)
     const baseCommand = `DOCKER_DEFAULT_PLATFORM=linux/amd64 docker run --rm -v \"${projectDir}\":/project -w /project ligolang/ligo:0.41.0 compile contract ${inputFile}`
     const entryPoint = opts.e ? `-e ${opts.e}` : ""
